\name{assortativity.phylo}
\alias{assortativity.phylo}
\title{
Compute assortativity coefficient from phylogenetic classifications
}
\usage{
assortativity.phylo(A, class, graph.unit, num.perm = 3000)
}
\arguments{
  \item{A}{pep adjacency matrix
}
  \item{class}{
 a data matrix that includes microbiome phylogenetic classification
}
  \item{graph.unit}{
the column name of the class matrix corresponding to the unit for the microbiome network
}
  \item{num.perm}{
number of permutation values}
}




