% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{class}
\alias{class}
\title{A data.frame of taxonomic information}
\format{
A data frame of taxonomic information with sample names in the row names.
}
\usage{
data("class")
}
\description{
This is hypothetically used but as there are no vignettes, examples, or tutorials it is unclear why this was included.
}
\keyword{datasets}
