% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{iOraldat}
\alias{iOraldat}
\title{An Oral microbiome dataset}
\format{
A count matrix
}
\usage{
data("iOraldat")
}
\description{
This is hypothetically used but as there are no vignettes, examples, or tutorials it is unclear why this was included.
}
\keyword{datasets}
