% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORadj.R
\name{ORadj}
\alias{ORadj}
\title{Turn a matrix of weights into a binary association matrix.}
\usage{
ORadj(mat)
}
\arguments{
\item{mat}{A square matrix.}
}
\value{
Returns a binary adjacency matrix
}
\description{
Turn a matrix of weights into a binary association matrix.
}
\examples{
set.seed(123)
x <- matrix(abs(round(rnorm(25, 0, 0.1), 1)), nrow=5)
x
ORadj(x)
}
