\name{perm.assortivity}
\alias{perm.assortivity}
\title{
Compute assortivity coefficient given a class variable with permutation
}
\usage{
perm.assortivity(G, x, perm = T, num.perm = 1000)
}
\arguments{
  \item{G}{
graph object
}
  \item{x}{
a vector for node label
}
  \item{perm}{
TRUE if permutation test is performed
}
  \item{num.perm}{
Number of permutation
}
}