% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assortativity.phylo.R
\name{assortativity.phylo}
\alias{assortativity.phylo}
\title{Compare assortivity of a network at various phylogenetic levels.}
\usage{
assortativity.phylo(A, class, graph.unit, num.perm = 3000)
}
\arguments{
\item{A}{An adjacency matrix for network}

\item{class}{A matrix of microbiome phylogenetic information. Note any NA values will all be categorized as the same.}

\item{graph.unit}{The column name of the class matrix that corresponds to the unit for the microbiome network. Columns before this one will be used.}

\item{num.perm}{A number of permutations to run, defaults to 3000. This is passed to \code{\link{perm.assortivity}}.}
}
\value{
Returns a list with two elements, "results" is a dataframe of results, "null.values" is a list of vectors showing null values for the permutation tests.
}
\description{
Compare assortivity of a network at various phylogenetic levels.
}
\examples{
x <- 4
}
\keyword{assortivity}
\keyword{permutation}
\keyword{phylogeny}
