% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.assortivity.R
\name{perm.assortivity}
\alias{perm.assortivity}
\title{Permutation test for a network node's assortivity}
\usage{
perm.assortivity(G, x, num.perm = 1000)
}
\arguments{
\item{G}{An igraph object}

\item{x}{Node labels from G as a numeric vector.}

\item{num.perm}{A number of permutations to run, defaults to 1000.}
}
\value{
Returns a list with two elements
The first element "result" is a datafram with two columns. "Value" is the observed nominal assortivity of node x, "p.value" is the percentage of permuted scores below that value.
The second element, "null.value", is the list of permuted assortivity scores.
}
\description{
Permutation test for a network node's assortivity
}
\examples{
x <- 4
}
\keyword{assortivity}
\keyword{permutation}
\keyword{phylogeny}
